code/python/flask_app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.6.2
boto3==1.34.154
botocore==1.34.154
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.12.14
charset-normalizer==3.3.2
click==8.1.7
datasets==2.19.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.13.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.3.1
gmpy2==2.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface==0.0.1
huggingface_hub==0.24.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
networkx==3.3
numexpr==2.10.1
numpy==1.26.4
openai==1.58.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.3.1
propcache==0.2.0
pyarrow==17.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
s3transfer==0.10.1
sacremoses==0.0.43
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
Werkzeug==3.0.6
wheel==0.44.0
xxhash==2.0.2
yarl==1.18.0
python-dotenv==1.0.1
